src/requirements.txt
bio==0.4.1
biopython==1.78
matplotlib==3.4.2
numpy==1.19.5
pandas==1.2.4
pillow==8.2.0
pyparsing==2.4.7
requests==2.25.1
seaborn==0.11.1
urllib3==1.26.4
bcbio-gff
dna_features_viewer
